Add files via upload
Fully updated to include latest code changes. Program is now working with the use of a number system for each part of the grid. Currently ties are nto restarting the program but this should work later. Will include a readme maybe for a full overview of the program as design changes are made.